Move Digest out of the digest module
